libparse: fix several implicit sign changes.

+++ libparse/clk_trimtaip.c
@@ -118,7 +118,8 @@ inp_trimtaip(
 {
 	unsigned int rtc;
 
-	parseprintf(DD_PARSE, ("inp_trimtaip(0x%lx, 0x%x, ...)\n", (long)parseio, ch));
+	parseprintf(DD_PARSE, ("inp_trimtaip(0x%lx, 0x%x, ...)\n",
+                    (unsigned long)parseio, (unsigned)ch));
 
 	switch (ch)
 	{
