+++ libparse/clk_varitext.c
@@ -160,7 +160,8 @@ inp_varitext(
   struct varitext *t = (struct varitext *)parseio->parse_pdata;
   int    rtc;
 
-  parseprintf(DD_PARSE, ("inp_varitext(0x%lx, 0x%x, ...)\n", (long)parseio, ch));
+  parseprintf(DD_PARSE, ("inp_varitext(0x%lx, 0x%x, ...)\n",
+              (unsigned long)parseio, (unsigned)ch));
 
   if (!t)
     return PARSE_INP_SKIP;	/* local data not allocated - sigh! */
